requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
gunicorn
Werkzeug